ai_service/ai_service_local/requirements.txt
fastapi
uvicorn[standard]
ultralytics
Pillow
python-multipart
numpy<2.0
python-dotenv
transformers
